pyproject.toml
[project]
name = "svg2pptx"
dynamic = ["version"]
description = "Convert SVG files to native, editable PowerPoint shapes"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "svg2pptx contributors"}
]
keywords = ["svg", "powerpoint", "pptx", "converter", "vector", "graphics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Office/Business :: Office Suites",
]

dependencies = [
    "python-pptx>=0.6.21",
    "svgpathtools>=1.6.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
svg2pptx = "svg2pptx.cli:main"

[project.urls]
Homepage = "https://github.com/svg2pptx/svg2pptx"
Repository = "https://github.com/svg2pptx/svg2pptx"

[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
# Automatically get version from git tags (e.g., v1.0.0 -> 1.0.0)
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/svg2pptx"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
]
